FInishi user controller

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -5,16 +5,21 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.dto.ListPageDto;
+import com.example.demo.dto.Token;
 import com.example.demo.dto.UserDto;
 import com.example.demo.dto.Web.CreateUserDto;
+import com.example.demo.dto.Web.MessageDto;
+import com.example.demo.dto.Web.UpdateUserDto;
 import com.example.demo.services.AuthService;
 import com.example.demo.services.UserService;
 
@@ -28,32 +33,36 @@ public class UserController {
     AuthService auth;
 
     @PostMapping("/user")
-    public ResponseEntity<String> createUser(@RequestBody CreateUserDto data){
+    public ResponseEntity<MessageDto> createUser(@RequestBody CreateUserDto data){
 
         if(!auth.validateEdvIsFree(data.EDV()))
-            return new ResponseEntity<>("EDV ja Cadastrado!",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageDto("EDV ja Cadastrado!"),HttpStatus.BAD_REQUEST);
 
         if(!auth.validateEdv(data.EDV()))
-            return new ResponseEntity<>("EDV invalido!",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageDto("EDV invalido!"),HttpStatus.BAD_REQUEST);
         
         if(!auth.validateEmailIsFree(data.email()))
-            return new ResponseEntity<>("Email  ja Cadastrado!",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageDto("Email  ja Cadastrado!"),HttpStatus.BAD_REQUEST);
 
         if(!auth.validateEmail(data.email()))
-            return new ResponseEntity<>("Email invalido!",HttpStatus.BAD_REQUEST); 
+            return new ResponseEntity<>(new MessageDto("Email invalido!"),HttpStatus.BAD_REQUEST); 
         
         if(!auth.ValidatePassword(data.password()))
-            return new ResponseEntity<>("Senha invalida!Deve conter Numero,Letra maiscula e miniscula",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageDto("Senha invalida!Deve conter Numero,Letra maiscula e miniscula"),HttpStatus.BAD_REQUEST);
 
         if(service.create(data.name(), data.EDV(), data.email(), data.password(),data.phone())==null)
-            return new ResponseEntity<>("Errou ao criar usuario",HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageDto("Errou ao criar usuario"),HttpStatus.BAD_REQUEST);
 
-        return new ResponseEntity<>("Usuario criado com sucesso!",HttpStatus.OK); 
+        return new ResponseEntity<>(new MessageDto("Usuario criado com sucesso!"),HttpStatus.OK); 
     }
 
     @GetMapping("/user")
-    public  ResponseEntity<ListPageDto<UserDto>> getAllUser(@RequestParam Integer page,@RequestParam Integer size){
-        return new ResponseEntity<>(service.getAll(page, size),HttpStatus.OK);
+    public  ResponseEntity<ListPageDto<UserDto>> getAllUser(@RequestParam(defaultValue = "0") Integer page,@RequestParam(defaultValue = "0") Integer size){
+        try {
+            return new ResponseEntity<>(service.getAll(page, size),HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping("/user/{id}")
@@ -65,15 +74,30 @@ public class UserController {
         return new ResponseEntity<>(user,HttpStatus.OK);
     }
 
+    @PutMapping("/user")
+    public ResponseEntity<MessageDto> updateUser(@RequestAttribute("token") Token token,@ModelAttribute UpdateUserDto data){
+        var res = service.update(token.getId(), data.name(), data.EDV(), data.email(),  null, data.phone(), false, data.student(), data.image(), null);
+        
+        if(!res.success())
+            return new ResponseEntity<>(new MessageDto(res.response()),HttpStatus.BAD_REQUEST);
+        
+        return new ResponseEntity<>(new MessageDto("Usuario atualizado com sucesso!"),HttpStatus.OK);
+    }
+
+
     @DeleteMapping("/admin/user/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id){
+    public ResponseEntity<MessageDto> deleteUser(@RequestAttribute("token") Token token,@PathVariable Long id){
+
+        if(!token.isAdmin()){
+            return new ResponseEntity<>(new MessageDto("Usuario sem autorização"),HttpStatus.UNAUTHORIZED);
+        }
 
         var res = service.delete(id);
 
         if(!res.success())
-            return new ResponseEntity<>(res.response(),HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageDto(res.response()),HttpStatus.BAD_REQUEST);
         
-        return new ResponseEntity<>("Usuario deletado com sucesso",HttpStatus.OK);
+        return new ResponseEntity<>(new MessageDto("Usuario deletado com sucesso"),HttpStatus.OK);
         
     }
